页面3:完成

+++ app/src/main/java/com/qxy/dousheng/model/InfoJson.kt
@@ -11,5 +11,5 @@ data class InfoJson(
 
 data class InfoExtra(
     var logid: String = "",
-    var now: Int = 0
+    var now: Long = 0
 )
